requirements.txt
Django~=5.1.4
djangorestframework~=3.15.2
markdown~=3.7
django-filter~=24.3
psycopg[binary]~=3.2.3
python-dotenv~=1.0.1